Cambios en Usuarios

+++ src/main/java/com/jdc/tallerhospital/repository/UsuariosRepository.java
@@ -3,5 +3,5 @@ package com.jdc.tallerhospital.repository;
 import com.jdc.tallerhospital.entity.UsuariosEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UsuariosRepository extends JpaRepository<UsuariosEntity, Long> {
+public interface UsuariosRepository extends JpaRepository<UsuariosEntity, Integer> {
 }
